Created MainData folder and for seperation of concern removed the data from Maincomponent & Menu component & placed it inside the MainData & MenuDat files
