+++ Components/Data/MainData.js
@@ -1,0 +1,34 @@
+const data = [
+  {
+    name: 'Stewed Mushrooms',
+    image: require('../assets/images/food1.jpg'),
+    rating: 3,
+    price: '$12',
+  },
+  {
+    name: 'Jackfruit Fried',
+    image: require('../assets/images/food2.jpg'),
+    rating: 5,
+    price: '$15',
+  },
+  {
+    name: 'Noodles',
+    image: require('../assets/images/food3.jpg'),
+    rating: 4,
+    price: '$20',
+  },
+  {
+    name: 'Beef',
+    image: require('../assets/images/food4.jpg'),
+    rating: 2,
+    price: '$12',
+  },
+  {
+    name: 'Salad dressing',
+    image: require('../assets/images/food5.jpg'),
+    rating: 5,
+    price: '$13',
+  },
+];
+
+export default data;
