+++ Components/MainMenu/Main/Main.js
@@ -8,44 +8,12 @@ import {
   TouchableOpacity,
 } from 'react-native';
 
+import data from '../../Data/MainData';
 import LinearGradient from 'react-native-linear-gradient';
 import Icon from 'react-native-vector-icons/Ionicons';
 import AntIcon from 'react-native-vector-icons/AntDesign';
 Icon.loadFont();
 AntIcon.loadFont();
-
-const data = [
-  {
-    name: 'Stewed Mushrooms',
-    image: require('../../assets/images/food1.jpg'),
-    rating: 3,
-    price: '$12',
-  },
-  {
-    name: 'Jackfruit Fried',
-    image: require('../../assets/images/food2.jpg'),
-    rating: 5,
-    price: '$15',
-  },
-  {
-    name: 'Noodles',
-    image: require('../../assets/images/food3.jpg'),
-    rating: 4,
-    price: '$20',
-  },
-  {
-    name: 'Beef',
-    image: require('../../assets/images/food4.jpg'),
-    rating: 2,
-    price: '$12',
-  },
-  {
-    name: 'Salad dressing',
-    image: require('../../assets/images/food5.jpg'),
-    rating: 5,
-    price: '$13',
-  },
-];
 
 class Main extends React.Component {
   state = {
