+++ Components/MainMenu/Menu/Menu.js
@@ -9,79 +9,7 @@ import {
   TouchableOpacity,
 } from 'react-native';
 
-const Data = [
-  {
-    type: 'Dessert',
-    color: '#f7932e',
-    data: [
-      {
-        name: 'Cream Sweet',
-        image: require('../../assets/images/creamsweet.jpg'),
-        price: '$12',
-      },
-      {
-        name: 'Cup Cakes',
-        image: require('../../assets/images/cupcakes.jpg'),
-        price: '$15',
-      },
-      {
-        name: 'Orange Cake',
-        image: require('../../assets/images/orangecake.jpg'),
-        price: '$10',
-      },
-      {
-        name: 'Strawberries',
-        image: require('../../assets/images/strawberries.jpg'),
-        price: '$18',
-      },
-    ],
-  },
-  {
-    type: 'Main course',
-    color: '#39b54a',
-    data: [
-      {
-        name: 'Spaghetti Noodles',
-        image: require('../../assets/images/spaghetti.jpg'),
-        rating: 4,
-        price: '$20',
-      },
-      {
-        name: 'Thai Noodles',
-        image: require('../../assets/images/thai.jpg'),
-        rating: 2,
-        price: '$12',
-      },
-      {
-        name: 'Ramen Noodles',
-        image: require('../../assets/images/ramen.jpg'),
-        rating: 5,
-        price: '$13',
-      },
-    ],
-  },
-  {
-    type: 'Other',
-    color: '#ed3e69',
-    data: [
-      {
-        name: 'Seafood Prawns',
-        image: require('../../assets/images/seafoodprawns.jpg'),
-        price: '$20',
-      },
-      {
-        name: 'Bread Salad',
-        image: require('../../assets/images/bread.jpg'),
-        price: '$13',
-      },
-      {
-        name: 'Chicken Feta',
-        image: require('../../assets/images/chickenfeta.jpg'),
-        price: '$15',
-      },
-    ],
-  },
-];
+import Data from '../../Data/MenuData';
 
 class Menu extends React.Component {
   state = {
